Wrong x/y recalculations, fixed.
Thanks to christopho, the conditions to remove the enemy are now more clear.
The function test_obstacles is relative to the entity, not the map, so the
calculations must be enemy_pos - test_pos, and not only test_pos. Same thing
for the set_xy. Next task : tests with this new parameters.