Clean & Test
Clean of a non useful line in gigas stopping timer on restart (it is
already done by the engine)

Mini test on the debugger enemy to check if it crashes the game when
veryfing the existance of the "master enemy" before to restart it.
Expectations : it will not crash, and can be considered as a good
solution until the engine fix this behavior.

Preparing a test function of the debugger enemy to check the
capabilities of the map and the engine when loading a lot, lot of mobs