Added a debug room and fix the thunder strike target
Still have to do tests about the delay of animation, but thunder strike
target is now OK. The second version will include the split on two strikes
at the same time.

Added a debug room, accessible from the hub with the konami code (up up
down down left right left right action). It's more for the fun than a real
debug room, but it can be used for whatever we need without having to
charge a full enemy or anything. Actually, it tests the possibility to
move the hero by the engine and to decompose thunder strike easily.