Set apple cure poison too
The apples were very useless in this quest. Now, they can cure poison (in
addition to give 1 luain). It is more convenient to set it to an existing
item than create a new item, completely review the pause menu to add it,
etc. Later, the slow and confusion will be cured probably Croissant and
Pain au Chocolat. Ideally, I would like to create spells to cure each
status and a master spell to cure all status, but it's not my most urgent
task.